Added a main menu test scene that also has a pause menu test on it
seems to work so far with no problems I can see. Main menu seems to load up a level and works going back to it as well. Pause menu stops unity's time system and starts it back up again when resumed as expected. Overall all the code seems to be working so far, which leads to a great success!