projectiles and movement working
the projectiles are currently working and the 3 speeds work via Boolean functions. Those being the normal upgrade, the fast upgrade and the fastest upgrade. So far everything seems to be working but I am going to work on this a bit more to make sure that everything is working to the desired operation. 👌