Bullet Pool System
Created a system for bullet pooling.
Scripts now have access to a pool of bullets to reference and whatever we decide to do with them.

When obtaining a bullet, select what type it should be (that is if we decide to have unique projectiles) and provide a scriptable object asset to give the bullet functionality.

Currently needs testing for enemy functionality.